refactor(type_checker): move CLI logic into lib, mirror typescript-go layout
Make `main.rs` a thin binary (like tsgo's `cmd/tsgo/main.go`) that just delegates to the library, and lay the modules out to match typescript-go: `execute/tsc.rs` (internal/execute/tsc.go) holds `command_line`/`tsc_compilation`/`find_config_file`, and `tsoptions/commandlineparser.rs` (internal/tsoptions/commandlineparser.go) holds the `bpaf` command and `parse_command_line`. No behavior change.
