+++ crates/oxc_type_checker/src/execute/mod.rs
@@ -1,0 +1,7 @@
+//! Port of typescript-go's `internal/execute` package.
+//!
+//! Drives the compiler from the command line.
+
+mod tsc;
+
+pub use tsc::command_line;
