+++ crates/oxc_type_checker/src/tsoptions/mod.rs
@@ -1,0 +1,7 @@
+//! Port of typescript-go's `internal/tsoptions` package.
+//!
+//! Command-line and (later) `tsconfig.json` option parsing.
+
+mod commandlineparser;
+
+pub use commandlineparser::{TypeCheckCommand, parse_command_line};
